make math easier to read; fix chipset int length; additions to CardScale
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@41561 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/private/graphics/radeon_hd/radeon_hd.h
@@ -106,7 +106,7 @@ struct radeon_shared_info {
 	uint16			cursor_hot_x;
 	uint16			cursor_hot_y;
 
-	uint32			device_chipset;
+	uint16			device_chipset;
 	char			device_identifier[32];
 	struct pll_info	pll_info;
 };
